+++ src/main/java/com/linkage/client/VendorService.java
@@ -5,7 +5,7 @@ import java.util.Map;
 
 import com.linkage.LinkageConfiguration;
 import com.linkage.api.ApiResponse;
-import com.linkage.core.validations.DigitapSchema.SendAadharOtp;
+import com.linkage.core.validations.VendorSchema.SendAadharOtp;
 
 import jakarta.ws.rs.core.MultivaluedHashMap;
